+++ just_audio/CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.10.5
+
+* Disable Android audio offload by default to prevent playback issues.
+* Add androidAudioOffloadPreferences.
+* Deprecate androidOffloadSchedulingEnabled.
+
+## 0.10.4
+
+* Fix bug on simultaneous loads.
+* Fix dispose crash on iOS.
+* Fix initial seek values on subsequent loads (@Abestanis).
+
+## 0.10.3
+
+* Fix pending timers bug in unit tests.
+* Fix NPE in Android position broadcast.
+
 ## 0.10.2
 
 * Add Player.playerEvent.